# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KisanDirect API — farm-to-village selling platform
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface CustomerAuthInput {
  phone: string;
  otp: string;
  name: string;
  village: string;
}

export interface Customer {
  id: number;
  name: string;
  phone: string;
  village: string;
  /** @nullable */
  created_at?: string | null;
}

export interface CustomerAuthResult {
  success: boolean;
  customer: Customer;
}

export interface SellerAuthInput {
  phone: string;
  otp: string;
}

export interface SellerAuthResult {
  success: boolean;
  /** @nullable */
  message?: string | null;
}

export interface Variety {
  id: number;
  product_id: number;
  name: string;
  price_per_kg: number;
  /** @nullable */
  description?: string | null;
  /** @nullable */
  shelf_life?: string | null;
  in_stock: boolean;
}

export interface Product {
  id: number;
  name: string;
  name_en: string;
  emoji: string;
  category: string;
  min_kg: number;
  bg_color: string;
  /** @nullable */
  created_at?: string | null;
  varieties?: Variety[];
  benefits?: string[];
}

export type ProductInputVarietiesItem = {
  name: string;
  price_per_kg: number;
  description?: string;
  shelf_life?: string;
  in_stock?: boolean;
};

export interface ProductInput {
  name: string;
  name_en: string;
  emoji: string;
  category: string;
  min_kg: number;
  bg_color: string;
  varieties?: ProductInputVarietiesItem[];
  benefits?: string[];
}

export interface StockUpdate {
  in_stock: boolean;
  /** @nullable */
  stock_level?: string | null;
}

export interface OrderItem {
  id: number;
  order_id: number;
  variety_id: number;
  product_name: string;
  variety_name: string;
  price_per_kg: number;
  quantity_kg: number;
}

export type OrderStatus = typeof OrderStatus[keyof typeof OrderStatus];


export const OrderStatus = {
  placed: 'placed',
  accepted: 'accepted',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export type OrderPaymentStatus = typeof OrderPaymentStatus[keyof typeof OrderPaymentStatus];


export const OrderPaymentStatus = {
  pending: 'pending',
  paid: 'paid',
} as const;

export interface Order {
  id: number;
  customer_id: number;
  village: string;
  /** @nullable */
  address?: string | null;
  total_amount: number;
  status: OrderStatus;
  payment_status: OrderPaymentStatus;
  return_requested?: boolean;
  /** @nullable */
  return_note?: string | null;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  customer_name?: string | null;
  /** @nullable */
  customer_phone?: string | null;
  items?: OrderItem[];
}

export type OrderInputItemsItem = {
  variety_id: number;
  product_name: string;
  variety_name: string;
  price_per_kg: number;
  quantity_kg: number;
};

export interface OrderInput {
  customer_id: number;
  village: string;
  address?: string;
  items: OrderInputItemsItem[];
}

export type OrderStatusUpdateStatus = typeof OrderStatusUpdateStatus[keyof typeof OrderStatusUpdateStatus];


export const OrderStatusUpdateStatus = {
  placed: 'placed',
  accepted: 'accepted',
  out_for_delivery: 'out_for_delivery',
  delivered: 'delivered',
  cancelled: 'cancelled',
} as const;

export interface OrderStatusUpdate {
  status: OrderStatusUpdateStatus;
}

export interface ReturnRequest {
  note: string;
}

export interface Village {
  id: number;
  name: string;
}

export type DashboardStatsStockSummaryItem = {
  product_name: string;
  variety_name: string;
  in_stock: boolean;
  /** @nullable */
  stock_level?: string | null;
};

export interface DashboardStats {
  new_orders: number;
  today_earnings: number;
  village_count: number;
  total_orders_today: number;
  low_stock_count: number;
  stock_summary?: DashboardStatsStockSummaryItem[];
}

export interface DeleteResult {
  success: boolean;
}

export type GetProductsParams = {
category?: string;
search?: string;
};

export type GetOrdersParams = {
phone?: string;
status?: string;
};