# lib/api-zod/src/generated/types/customerAuthInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KisanDirect API — farm-to-village selling platform
 * OpenAPI spec version: 0.1.0
 */

export interface CustomerAuthInput {
  phone: string;
  otp: string;
  name: string;
  village: string;
}