# lib/api-zod/src/generated/types/dashboardStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KisanDirect API — farm-to-village selling platform
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardStatsStockSummaryItem } from './dashboardStatsStockSummaryItem';

export interface DashboardStats {
  new_orders: number;
  today_earnings: number;
  village_count: number;
  total_orders_today: number;
  low_stock_count: number;
  stock_summary?: DashboardStatsStockSummaryItem[];
}