# lib/api-zod/src/generated/types/order.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KisanDirect API — farm-to-village selling platform
 * OpenAPI spec version: 0.1.0
 */
import type { OrderItem } from './orderItem';
import type { OrderPaymentStatus } from './orderPaymentStatus';
import type { OrderStatus } from './orderStatus';

export interface Order {
  id: number;
  customer_id: number;
  village: string;
  /** @nullable */
  address?: string | null;
  total_amount: number;
  status: OrderStatus;
  payment_status: OrderPaymentStatus;
  return_requested?: boolean;
  /** @nullable */
  return_note?: string | null;
  /** @nullable */
  created_at?: string | null;
  /** @nullable */
  customer_name?: string | null;
  /** @nullable */
  customer_phone?: string | null;
  items?: OrderItem[];
}