# lib/api-zod/src/generated/types/orderInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * KisanDirect API — farm-to-village selling platform
 * OpenAPI spec version: 0.1.0
 */
import type { OrderInputItemsItem } from './orderInputItemsItem';

export interface OrderInput {
  customer_id: number;
  village: string;
  address?: string;
  items: OrderInputItemsItem[];
}